Dedup redundant type/kind info by simplifying full signature generation
For completions that don't have a signature, `BaseName.description`
provides type, name, and typing information if it was explicitly
written.  For completions that provide signatures, `.to_string` provides
the parameters and typing information that needs to be appended to the
type.

`get_type_hints()` is an expensive call that calls into `jedi.infer()`,
even for completions that have explicit typing information.